feat: :sparkles: created whole application for parking lot with one floor

+++ parking/Parking.java
@@ -4,5 +4,5 @@ import exception.ParkingFullException;
 public interface Parking {
     public Ticket park(Vehicle vehicle) throws ParkingFullException;
 
-    public int unPark(Ticket ticket, ParkingChargeStretegy ParkingChargeStretegy) throws InvalidVehicleNumberException;
+    public long unPark(Ticket ticket, ParkingChargeStretegy ParkingChargeStretegy) throws InvalidVehicleNumberException;
 }
